Update: remove `docopt` dependency in agent
Running the agent with `python3 start.py` is sufficient.
The agent doesn't accept any CLI args, so `docopt` dependency is useless.
