+++ agent/.dockerignore
@@ -14,4 +14,5 @@ __pycache__/
 
 # other
 .flake8
-test/
+.env
+docker-compose.yml
