--- agent/dnsagent/clis/__init__.py
@@ -1,1 +1,0 @@
-from .start import *
